rename git.ts to cmds.ts
Rename this file with the intent to move all the command callouts to one
place.
